+++ src/operators/frei0r_op.ml
@@ -23,6 +23,8 @@
 open Source
 open Extralib
 
+let () = Lang.add_module "video.frei0r"
+
 type t = Float | Int | Bool
 
 let log = Log.make ["frei0r"]