+++ src/operators/iir_filter.ml
@@ -458,6 +458,9 @@ class iir ~kind (source : source) filter_family filter_type order freq1 freq2
   end
 
 let () =
+  Lang.add_module "filter.iir";
+  Lang.add_module "filter.iir.butterworth";
+  Lang.add_module "filter.iir.resonator";
   let kind = Lang.any in
   let k = Lang.kind_type_of_kind_format kind in
   Lang.add_operator "filter.iir.butterworth.high"